fix: Suspense 코드 추가로 인한 SSR 오류 해결
